requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
python-dotenv
Werkzeug
Flask-JWT-Extended
requests